backend/requirements.txt
# Core FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image processing and AI
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
numpy>=1.24.0

# Computer vision (optional but useful)
opencv-python>=4.8.0
easyocr>=1.7.0
ultralytics>=8.0.0

# HTTP and API
requests>=2.31.0

# Configuration and validation
python-dotenv>=1.0.0
pydantic>=2.0.0,<3.0.0

# Text processing
ftfy>=6.1.0
regex>=2023.0.0

# Google AI (for Gemini API)
google-generativeai>=0.3.0

# Development and testing (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0

# Note: CLIP installation from GitHub (optional)
# git+https://github.com/openai/CLIP.git
